Refactoring navigation
