+++ babel.config.js
@@ -1,6 +1,7 @@
 module.exports = function(api) {
-  api.cache(true);
+  api.cache(true)
+
   return {
     presets: ['babel-preset-expo'],
-  };
-};
+  }
+}
