+++ src/components/common/navigation/BottomTabNavigator.js
@@ -1,19 +1,16 @@
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import * as React from "react";
 
-import TabBarIcon from "../components/TabBarIcon";
-import Explore from "../screens/Explore";
-import Podcasts from "../screens/Podcasts";
-import Social from "../screens/Social";
-import TopCharts from "../screens/TopCharts";
+import TabBarIcon from "../../../../components/TabBarIcon";
+import Explore from "../../screens/Explore";
+import Podcasts from "../../screens/Podcasts";
+import Social from "../../screens/Social";
+import TopCharts from "../../screens/TopCharts";
 
 const BottomTab = createBottomTabNavigator();
 const INITIAL_ROUTE_NAME = "Podcasts";
 
 export default function BottomTabNavigator({ navigation, route }) {
-  // Set the header title on the parent stack navigator depending on the
-  // currently active tab. Learn more in the documentation:
-  // https://reactnavigation.org/docs/en/screen-options-resolution.html
   navigation.setOptions({ headerTitle: getHeaderTitle(route) });
 
   return (
